requirements.txt
flask
nemo_toolkit[all]==1.22.0
soundfile
requests
